Allow for explicit header order
Defaulting to alphabetized header order is a good start, but some
times (and for some old and terrible systems) you have to be able
to ensure that headers appear in a pre-determined order.

By making Header ([]string) globally accessible, we can accomplish
this task.

Note that if you manually set the Header attribute, it will ignore
any attributes passed in that are not set within the Header slice.

This can be for good or ill, but it is a thing that could bite you
if you weren't expecting it.